Allow for better filtering of tests
